3.8/n move run_attribution_example into docs_snippets
