+++ docs/content/guides/dagster/run-attribution.mdx
@@ -5,13 +5,11 @@ description: How to perform Run Attribution by using the Run Coordinator.
 
 # Run Attribution <Experimental />
 
-<CodeReferenceLink filePath="examples/run_attribution_example" />
-
 ## Run Coordinator
 
-The <Link href="/deployment/run-coordinator">Run Coordinator</Link> is used to control the policy that Dagster uses to manage the set of active runs on your deployment.
+The [Run Coordinator](/deployment/run-coordinator) is used to control the policy that Dagster uses to manage the set of active runs on your deployment.
 
-The Run Coordinator is invoked when runs are submitted on the <Link href="/deployment/dagster-instance">instance</Link> (e.g. via the <Link href="/concepts/dagit/graphql#launch-a-pipeline-run">GraphQL API</Link>), and as a result it can be used to dynamically attach <Link href="/concepts/solids-pipelines/pipelines#pipeline-tags">tags</Link> to submitted runs.
+The Run Coordinator is invoked when runs are submitted on the [instance](/deployment/dagster-instance) (e.g. via the [GraphQL API](/concepts/dagit/graphql#launch-a-pipeline-run)), and as a result it can be used to dynamically attach <Link href="/concepts/solids-pipelines/pipelines#pipeline-tags">tags</Link> to submitted runs.
 
 In this example, we'll perform run attribution, which means that we'll attach a user's email as a tag to submitted runs.
 
@@ -19,9 +17,9 @@ To accomplish this, we'll use a custom Run Coordinator to read Flask HTTP header
 
 ### Custom Run Coordinator
 
-In this use case, we'd like to add a hook to customize submitted runs while still using a queue to submit runs to the [Dagster Daemon](/deployment/dagster-daemon). To accomplish this, we can use the <Link href="/deployment/run-coordinator#queuedruncoordinator">Queued Run Coordinator</Link> in the example below. The `context` object available in `submit_run` has a `get_request_header` method we can use to read HTTP headers:
+In this use case, we'd like to add a hook to customize submitted runs while still using a queue to submit runs to the [Dagster Daemon](/deployment/dagster-daemon). To accomplish this, we can use the [Queued Run Coordinator](/deployment/run-coordinator#queuedruncoordinator) in the example below. The `context` object available in `submit_run` has a `get_request_header` method we can use to read HTTP headers:
 
-```python file=/guides/dagster/run-attribution/custom_run_coordinator_skeleton.py startafter=start_custom_run_coordinator_marker endbefore=end_custom_run_coordinator_marker
+```python file=/guides/dagster/run_attribution/custom_run_coordinator_skeleton.py startafter=start_custom_run_coordinator_marker endbefore=end_custom_run_coordinator_marker
 from dagster._core.run_coordinator import QueuedRunCoordinator, SubmitRunContext
 from dagster._core.storage.pipeline_run import PipelineRun
 
@@ -33,7 +31,7 @@ class CustomRunCoordinator(QueuedRunCoordinator):
 
 Then we can parse the relevant header (in this case, called the `jwt_claims_header`) with any custom hook. In the following example, we're decoding a JWT header which contains the user's email.
 
-```python file=../../run_attribution_example/run_attribution_example/custom_run_coordinator.py startafter=start_email_marker endbefore=end_email_marker dedent=4
+```python file=/guides/dagster/run_attribution/custom_run_coordinator.py startafter=start_email_marker endbefore=end_email_marker dedent=4
 def get_email(self, jwt_claims_header: Optional[str]) -> Optional[str]:
     if not jwt_claims_header:
         return None
@@ -54,7 +52,7 @@ The above is just an example - you can write any hook which would be useful to y
 
 Putting this all together, we can use these hooks to dynamically attach tags to submitted pipeline runs. In the following example, we'd read the user's email from the `X-Amzn-Oidc-Data` header by using the `get_email` hook defined above, and then attach the email as a tag to the pipeline run.
 
-```python file=../../run_attribution_example/run_attribution_example/custom_run_coordinator.py startafter=start_submit_marker endbefore=end_submit_marker dedent=4
+```python file=/guides/dagster/run_attribution/custom_run_coordinator.py startafter=start_submit_marker endbefore=end_submit_marker dedent=4
 def submit_run(self, context: SubmitRunContext) -> PipelineRun:
     pipeline_run = context.pipeline_run
     jwt_claims_header = context.get_request_header("X-Amzn-Oidc-Data")
@@ -72,7 +70,7 @@ def submit_run(self, context: SubmitRunContext) -> PipelineRun:
 
 To specify the custom Run Coordinator to be used on the instance, add the following snippet to an instance's `dagster.yaml`:
 
-```yaml file=/guides/dagster/run-attribution/dagster.yaml
+```yaml file=/guides/dagster/run_attribution/dagster.yaml
 run_coordinator:
   module: run_attribution_example
   class: CustomRunCoordinator
@@ -82,7 +80,7 @@ run_coordinator:
 
 If you're using Helm to deploy instead, you can specify the custom run coordinator in the Helm chart's `values.yaml`:
 
-```yaml file=/guides/dagster/run-attribution/values.yaml
+```yaml file=/guides/dagster/run_attribution/values.yaml
 dagsterDaemon:
   runCoordinator:
     enabled: true
