--- examples/run_attribution_example/tox.ini
@@ -1,26 +1,0 @@
-[tox]
-envlist = py{39,38,37,36}-{unix,windows},mypy,pylint
-skipsdist = True
-
-[testenv]
-usedevelop = true
-extras =
-  test
-setenv =
-  VIRTUALENV_PIP=21.3.1
-passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE
-deps =
-  -e ../../python_modules/dagster[mypy,test]
-allowlist_externals =
-  /bin/bash
-commands =
-  !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
-  pytest -vv {posargs}
-
-[testenv:mypy]
-commands =
-  mypy --config=../../pyproject.toml --non-interactive --install-types {posargs} .
-
-[testenv:pylint]
-commands =
-  pylint -j0 --rcfile=../../pyproject.toml {posargs} run_attribution_example run_attribution_example_tests
